Update TableRendererTest et al to use new way of doing line width
