+++ cdap-cli/src/test/java/co/cask/cdap/cli/util/table/AltStyleTableRendererTest.java
@@ -15,17 +15,51 @@
  */
 package co.cask.cdap.cli.util.table;
 
+import com.google.common.base.Strings;
+import org.junit.FixMethodOrder;
 import org.junit.Ignore;
+import org.junit.Test;
+import org.junit.runners.MethodSorters;
 
 /**
  *
  */
 @Ignore
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class AltStyleTableRendererTest extends TableRendererTest {
 
   @Override
   public TableRenderer getRenderer() {
     return new AltStyleTableRenderer();
   }
-  
+
+  @Test
+  public void testOneCharLongerThanWidth() {
+    Table table = Table.builder()
+      .setHeader("c1")
+      .setRows(Table.rows()
+                 .add(Strings.repeat("z", (LINE_WIDTH - 4)) + "a")
+                 .build())
+      .build();
+    getRenderer().render(TEST_CONFIG, OUTPUT, table);
+  }
+
+  @Test
+  public void testHeaderOneCharLongerThanWidth() {
+    Table table = Table.builder()
+      .setHeader(Strings.repeat("z", (LINE_WIDTH - 4)) + "a")
+      .setRows(Table.rows().add("abc").build())
+      .build();
+    getRenderer().render(TEST_CONFIG, OUTPUT, table);
+  }
+
+  @Test
+  public void testOneCharLongerThan2Width() {
+    Table table = Table.builder()
+      .setHeader("c1")
+      .setRows(Table.rows().add(Strings.repeat("z", (LINE_WIDTH - 4) * 2) + "a").build())
+      .build();
+    getRenderer().render(TEST_CONFIG, OUTPUT, table);
+  }
+
 }
